Refactored class RepeatDay
